# contract/openapi.yaml
openapi: 3.0.0
info:
  title: Kafbat UI Api Service
  contact:
    name: API Support
    email: support@kafbat.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.2.0
...
paths:
...
  /api/clusters/{clusterName}/brokers:
    get:
      operationId: getBrokers
      summary: getBrokers
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Broker'
      tags:
        - Brokers
  /api/clusters/{clusterName}/brokers/csv:
    get:
      operationId: getBrokersCsv
      summary: getBrokersCsv
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            text/csv:
              schema:
                type: string
      tags:
        - Brokers
  /api/clusters/{clusterName}/brokers/logdirs:
    get:
      operationId: getAllBrokersLogdirs
      summary: getAllBrokersLogdirs
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: broker
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BrokersLogdirs'
      tags:
        - Brokers
  /api/clusters/{clusterName}/brokers/{id}/configs:
    get:
      operationId: getBrokerConfig
      summary: getBrokerConfig
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BrokerConfig'
      tags:
        - Brokers
  /api/clusters/{clusterName}/brokers/{id}/configs/{name}:
    put:
      operationId: updateBrokerConfigByName
      summary: updateBrokerConfigByName
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Brokers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerConfigItem'
  /api/clusters/{clusterName}/brokers/{id}/logdirs:
    patch:
      operationId: updateBrokerTopicPartitionLogDir
      summary: updateBrokerTopicPartitionLogDir
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Brokers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerLogdirUpdate'
...
  /api/clusters/{clusterName}/stats:
    get:
      operationId: getClusterStats
      summary: getClusterStats
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterStats'
      tags:
        - Clusters
...
components:
  schemas:
...
    Broker:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
        host:
          type: string
        port:
          type: integer
          format: int32
        bytesInPerSec:
          type: number
          format: double
        bytesOutPerSec:
          type: number
          format: double
        partitionsLeader:
          type: integer
          format: int32
        partitions:
          type: integer
          format: int32
        inSyncPartitions:
          type: integer
          format: int32
        partitionsSkew:
          type: number
          format: double
        leadersSkew:
          type: number
          format: double
    BrokerConfig:
      type: object
      required:
        - name
        - value
        - source
        - isSensitive
        - isReadOnly
      properties:
        name:
          type: string
        value:
          type: string
        source:
          $ref: '#/components/schemas/ConfigSource'
        isSensitive:
          type: boolean
        isReadOnly:
          type: boolean
        synonyms:
          type: array
          items:
            $ref: '#/components/schemas/ConfigSynonym'
    BrokerConfigItem:
      type: object
      properties:
        value:
          type: string
    BrokerDiskUsage:
      type: object
      required:
        - brokerId
      properties:
        brokerId:
          type: integer
          format: int32
        segmentSize:
          type: integer
          format: int64
        segmentCount:
          type: integer
          format: int32
    BrokerLogdirUpdate:
      type: object
      properties:
        topic:
          type: string
        partition:
          type: integer
          format: int32
        logDir:
          type: string
...
    BrokerTopicLogdirs:
      type: object
      properties:
        name:
          type: string
        partitions:
          type: array
          items:
            $ref: '#/components/schemas/BrokerTopicPartitionLogdir'
    BrokerTopicPartitionLogdir:
      type: object
      properties:
        partition:
          type: integer
          format: int32
        size:
          type: integer
          format: int64
        offsetLag:
          type: integer
          format: int64
        broker:
          type: integer
          format: int32
    BrokersLogdirs:
      type: object
      properties:
        name:
          type: string
        error:
          type: string
        topics:
          type: array
          items:
            $ref: '#/components/schemas/BrokerTopicLogdirs'
...
    ClusterStats:
      type: object
      properties:
        brokerCount:
          type: integer
          format: int32
        zooKeeperStatus:
          type: integer
          format: int32
          deprecated: true
        activeControllers:
          type: integer
          format: int32
          description: Id of broker which is cluster's controller. null, if controller not known yet.
        onlinePartitionCount:
          type: integer
          format: int32
        offlinePartitionCount:
          type: integer
          format: int32
        inSyncReplicasCount:
          type: integer
          format: int32
        outOfSyncReplicasCount:
          type: integer
          format: int32
        underReplicatedPartitionCount:
          type: integer
          format: int32
        diskUsage:
          type: array
          items:
            $ref: '#/components/schemas/BrokerDiskUsage'
        version:
          type: string
...
    ConfigSource:
      type: string
      enum:
        - DYNAMIC_TOPIC_CONFIG
        - DYNAMIC_BROKER_LOGGER_CONFIG
        - DYNAMIC_BROKER_CONFIG
        - DYNAMIC_DEFAULT_BROKER_CONFIG
        - DYNAMIC_CLIENT_METRICS_CONFIG
        - STATIC_BROKER_CONFIG
        - DEFAULT_CONFIG
        - UNKNOWN
    ConfigSynonym:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        source:
          $ref: '#/components/schemas/ConfigSource'
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
        - requestId
      properties:
        code:
          type: integer
          format: int32
          description: Internal error code (can be used for message formatting & localization on UI)
        message:
          type: string
          description: Error message
        timestamp:
          type: number
          format: decimal
          description: Response unix timestamp in ms
        requestId:
          type: string
          description: Unique server-defined request id for convenient debugging
        fieldsErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        stackTrace:
          type: string
    FieldError:
      type: object
      required:
        - fieldName
      properties:
        fieldName:
          type: string
          description: Name of field that violated format
        restrictions:
          type: array
          items:
            type: string
          description: Field format violations description (ex. ["size must be between 0 and 20", "must be a well-formed email address"])
...
servers:
  - url: http://localhost:8080
    description: Default endpoint for Kafbat UI API
    variables: {}